Refactor widget instrumentation tests so they are independent of other activities and instrumentation tests.
- Create a new activity used specifically for motion photo widget tests
- Update Android manifest and new activity layout
- Update main activity to separate it from motion photo widget
- Use new activity in MotionPhotoWidgetTest
- Fix null reader calls in MotionPhotoReader
